failed attempt to make windows work
I figured if both allonet and lovr used lua51 instead of
luajit, there'd be a lua.dll in the app directory and both
allonet and lovr could use it. BUT then cmake has duplicate
targets called "liblua" because allonet and lover have their
own stupid copies of lua. So this doesn't work. Giving up
and moving to GetProcAddress, just committing this
for posterity.