Use keyboard IJKLUO to joystick-move
WASD is currently used to move the virtual headset around in the world. In addition, you should also be able to walk around beyond your guardian bounds, so you walk with your left and right stick.

When using the desktop visor, you don't have sticks. So, these are now mapped IJKL for lateral movement and UO to snap-turn.